moving AlertBar to general components folder
